refactor: 拆分 Gateway dispatcher 会话 worker 状态机 (#114)
